move 3D poitn trait to types, small refactor of some functions in types with 3D point
